+++ src/user-py/src/capis.h
@@ -1,0 +1,11 @@
+#ifndef SKULLPY_C_APIS_H
+#define SKULLPY_C_APIS_H
+
+namespace skullpy {
+
+void register_capis();
+
+} // End of namespace
+
+#endif
+