+++ src/user-py/src/txn_idldata.h
@@ -1,0 +1,29 @@
+#ifndef SKULLPY_IDL_INTERNAL_H
+#define SKULLPY_IDL_INTERNAL_H
+
+#include <stddef.h>
+
+namespace skullpy {
+
+class TxnSharedRawData {
+private:
+    void*  data_;
+    size_t size_;
+
+public:
+    TxnSharedRawData() : data_(NULL), size_(0) {};
+    TxnSharedRawData(const void* data, size_t sz);
+    ~TxnSharedRawData();
+
+public:
+    const void* data() const;
+    void* data();
+    void  reset(const void* data, size_t sz);
+
+    size_t size() const;
+};
+
+}
+
+#endif
+